Make connstat more flexible with respect to what kernel it's running on.
Fixes #23 /proc/net/stats_tcp pads all lines with spaces up to 150 columns
Fixes #24 timestamp output is corrupt

This change adds support for deprecating a field. If/when this is ever
done, the connstat user-land and kernel packages will be updated
together, but until the OS is rebooted, the kernel will continue to
output a deprecated field that connstat would no longer need to know
about.

As part of this change, a connstat command unit test framework is
introduced along with a comprehensive set of unit tests that are
automatically run in pull-request and push context on GitHub via a new
GitHub action.

The bugs referenced above were found and fixed in the process of writing
these unit tests.